refac: mypage 리팩토링
